finish exercise 4.21-4.22

+++ blog-list/controllers/blogs.js
@@ -1,7 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+// const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
@@ -10,12 +10,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     if (!body.title || !body.url) {
         return response.status(400).json({
@@ -39,8 +34,27 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(204).end()
+    }
+
+    const user = await User.findById(request.user.id)
+    if (!user) {
+        return response.status(401).json({ error: 'token invalid: user not found' })
+    }
+
+    if (blog.user.toString() === user.id.toString() ) {
+        await Blog.findByIdAndDelete(request.params.id)
+
+        user.blogs = user.blogs.filter(
+            blogId => blogId.toString() !== request.params.id
+        )
+        await user.save()
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'token invalid: unauthorized to delete this blog' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
